requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Validation & Security
validators==0.22.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Development Tools
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# LLM Integration - Production Ready
google-generativeai>=0.3.0
anthropic>=0.8.0
tenacity>=8.2.0
tiktoken>=0.5.0